Audit obvious WIZ8 source-model mismatches

+++ include/wiz8/engine_code/materials.h
@@ -104,7 +104,7 @@ extern W8OctPreTreeVertex* g_gd_vertices_0065bd34;
 extern W8OctRegionPolygon* g_gd_polygons_0065bd38;
 
 char BuildPreprocessedFiles00492E60(const char* level_path);
-void ReportBuildStatus00497690(int channel, const char* message);
+void ReportBuildStatus00497690(short channel, const char* message);
 void ReportStartupMessage004969D0(const char* message);
 char* TrimAndLowercaseString(char* text);
 struct W8OctPreTreeVertex;
